+++ source/core/data/src/main/kotlin/com/xayah/core/data/repository/CloudRepository.kt
@@ -11,6 +11,7 @@ import com.xayah.core.rootservice.service.RemoteRootService
 import com.xayah.core.util.LogUtil
 import com.xayah.core.util.PathUtil
 import com.xayah.core.util.model.ShellResult
+import com.xayah.core.database.dao.UploadIdDao
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.first
@@ -22,6 +23,7 @@ class CloudRepository @Inject constructor(
     @ApplicationContext private val context: Context,
     private val rootService: RemoteRootService,
     private val cloudDao: CloudDao,
+    private val uploadIdDao: UploadIdDao
 ) {
     private fun log(msg: () -> String): String = run {
         LogUtil.log { "CloudRepository" to msg() }
@@ -132,7 +134,7 @@ class CloudRepository @Inject constructor(
     suspend fun getClient(name: String? = null): Pair<CloudClient, CloudEntity> {
         val entity = queryByName(name ?: context.readCloudActivatedAccountName().first())
         if (entity != null) if (entity.remote.isEmpty()) throw IllegalAccessException("${entity.name}: Remote directory is not set.")
-        val client = entity?.getCloud()?.apply { connect() } ?: throw NullPointerException("Client is null.")
+        val client = entity?.getCloud(uploadIdDao)?.apply { connect() } ?: throw NullPointerException("Client is null.")
         return client to entity
     }
 
@@ -150,7 +152,7 @@ class CloudRepository @Inject constructor(
         val clients: MutableList<Pair<CloudClient, CloudEntity>> = mutableListOf()
         cloudDao.queryActivated().forEach {
             if (it.remote.isEmpty()) throw IllegalAccessException("${it.name}: Remote directory is not set.")
-            clients.add(it.getCloud().apply { connect() } to it)
+            clients.add(it.getCloud(uploadIdDao).apply { connect() } to it)
         }
         block(clients)
         clients.forEach { it.first.disconnect() }