+++ source/core/network/src/main/kotlin/com/xayah/core/network/client/CloudClient.kt
@@ -11,6 +11,7 @@ import com.xayah.core.network.util.getExtraEntity
 import com.xayah.core.rootservice.parcelables.PathParcelable
 import com.xayah.libpickyou.parcelables.DirChildrenParcelable
 import com.xayah.core.model.database.S3Extra
+import com.xayah.core.database.dao.UploadIdDao
 
 interface CloudClient {
     fun connect()
@@ -36,7 +37,7 @@ interface CloudClient {
     suspend fun setRemote(context: Context, onSet: suspend (remote: String, extra: String) -> Unit)
 }
 
-fun CloudEntity.getCloud() = when (this.type) {
+fun CloudEntity.getCloud(uploadIdDao: UploadIdDao) = when (this.type) {
     CloudType.FTP -> {
         val extra = getExtraEntity<FTPExtra>()!!
         FTPClientImpl(this, extra)
@@ -59,6 +60,6 @@ fun CloudEntity.getCloud() = when (this.type) {
 
     CloudType.S3 -> {
         val extra = getExtraEntity<S3Extra>()!!
-        S3ClientImpl(this, extra)
+        S3ClientImpl(this, extra, uploadIdDao)  // 传递 uploadIdDao
     }
 }