+++ source/feature/main/processing/src/main/kotlin/com/xayah/feature/main/processing/medium/restore/RestoreViewModelImpl.kt
@@ -96,7 +96,7 @@ class RestoreViewModelImpl @Inject constructor(
                         )
                     )
                     runCatching {
-                        val client = state.cloudEntity!!.getCloud()
+                        val (client, _) = mCloudRepo.getClient(state.cloudEntity!!.name)
                         client.testConnection()
                         emitEffect(IndexUiEffect.DismissSnackbar)
                         withMainContext {