+++ source/feature/main/processing/src/main/kotlin/com/xayah/feature/main/processing/packages/backup/BackupViewModelImpl.kt
@@ -77,7 +77,7 @@ class BackupViewModelImpl @Inject constructor(
                         )
                     )
                     runCatching {
-                        val client = state.cloudEntity!!.getCloud()
+                        val (client, _) = mCloudRepo.getClient(state.cloudEntity!!.name)
                         client.testConnection()
                         emitEffect(IndexUiEffect.DismissSnackbar)
                         withMainContext {